Factor our initResource
We don't currently use different base urls for comms, so simplify the
usage somewhat. This allows us to extract out the init + optionally ask
for credentials logic, making the code easier to read.
